requirements.txt
Flask
requests
google-cloud-datastore==2.15.1